requirements.txt
streamlit>=1.35.0
opencv-python>=4.10.0
insightface==0.7.3
mediapipe==0.10.14
ultralytics>=8.1.34
pygame>=2.5.2
psycopg2-binary>=2.9.9
numpy>=1.26.4
pandas>=2.2.2
